chore: remove auction view controller
